Refactor: 네이밍 규칙따라 수정
