+++ src/main/java/com/kuit/chatdiary/service/DiaryService.java
@@ -1,6 +1,6 @@
 package com.kuit.chatdiary.service;
 
-import com.kuit.chatdiary.dto.GetDiaryDetailResponse;
+import com.kuit.chatdiary.dto.diary.DiaryShowDetailResponse;
 import com.kuit.chatdiary.repository.DiaryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,9 +20,7 @@ public class DiaryService {
     private final DiaryRepository diaryRepository;
 
     @Transactional
-
-
-    public List<GetDiaryDetailResponse> showDiary(Long userId, Date diaryDate) throws ParseException {
+    public List<DiaryShowDetailResponse> showDiary(Long userId, Date diaryDate) throws ParseException {
         log.info("[DiaryService.showDiary]");
         return diaryRepository.showDiaryDetail(userId, diaryDate);
     }
